requirements.txt
python-telegram-bot
requests
peewee
validators
pyquery
lxml
jsonpath-ng
dnspython